[FIX] imports in tests & refactoring
