+++ examples/data_analysis.py
@@ -6,8 +6,8 @@ from matplotlib import pyplot as plt
 from gsmmutils import DATA_PATH
 
 pd.set_option('display.max_columns', None)
-from gsmmutils.experimental.Biomass import Biomass
-from gsmmutils.experimental.ExpMatrix import ExpMatrix
+from gsmmutils.experimental.biomass import Biomass
+from gsmmutils.experimental.exp_matrix import ExpMatrix
 from gsmmutils.graphics.plot import boxplot, qqplot
 from gsmmutils.stats.stats import StatisticalAnalysis
