+++ src/gsmmutils/experimental/__init__.py
@@ -1,0 +1,1 @@
+from .biomass import Biomass
