+++ src/gsmmutils/experimental/biomass_component.py
